test(mirroring): merge mirror_shared_test.go into mirror_reconciler_test.go
